@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pipy-awc"
-version = "0.8.0"
+version = "0.9.0"
 description = "PiPy Automatic-Water Controller: A dynamically-configurable & automated aquarium water level controller for the Raspberry Pi using Python. Intended to run using mechanical voltage relays (i.e. Waveshare Relay Hat) and optical contact liquid sensors (i.e. from DFRobot/CQRObot)"
 classifiers = [
     "Development Status :: 3 - Alpha",
@@ -19,18 +19,17 @@ classifiers = [
     "Topic :: System :: Hardware",
 ]
 keywords = ["Raspberry Pi", "Aquarium", "Automation", "AWC", "RPi"]
-authors = [
-	{name = "Neil Kuehnle", email = "nkuehnle1191@gmail.com"}
-]
-license = {text = "GPLv2"}
+authors = [{ name = "Neil Kuehnle" }]
+license = { text = "GPLv2" }
 readme = "README.md"
-requires-python = "~=3.7"
+requires-python = ">=3.11"
 dependencies = [
-	"pandas>=1.3",
-	"pyaml>=21.10",
-	"schedule>=1.1.0",
-	"gpiozero>=1.6.2",
-	"imap-tools==0.55.0"
+    "apprise>=1.7.4",
+    "pyaml>=21.10",
+    "schedule>=1.1.0",
+    "gpiozero>=1.6.2",
+    "imap-tools>=1.5.0",
+    "python-dateutil>=2.90",
 ]
 
 [project.urls]
