@@ -1,77 +1,68 @@
-# Default module imports
-import datetime as dt
-from pathlib import Path
-from typing import Dict, List, Union, Any, Optional
+import inspect
+from typing import Dict, List, Optional
 
-# Third-party module imports
-import pandas as pd
-
-# Custom module imports
-# Control high-lvl logic, schedule, user input & more
-from .operations import (
-    Messenger,
-    AScheduler,
-    RemoteCommand,
+from .logistics import Routine
+from .peripherals import Dispenser, Monitor  # Interact w/ RPi GPIO headers/equipment
+from .services import (  # Control high-lvl logic, schedule, user input & more
+    AdvJob,
+    AdvScheduler,
     CancelJob,
+    Messenger,
+    NotificationFailure,
+    NotifyType,
+    RemoteCommand,
 )
-from .operations import MessengerError
-
-# Interact w/ RPi GPIO headers/equipment
-from .peripherals import (
-    Monitor,
-    Routine,
-    Dispenser,
-)
-from .peripherals import ErrorSensorTriggered, PumpTimeoutError
 
 TIME_FMT = "%m/%d/%Y: %H:%M:%S"
 
-HOME_DIR = Path(__file__).parents[1]
-LOG_DIR = HOME_DIR / "data" / "logs"
-try:
-    LOG_DIR.mkdir(parents=True, exist_ok=False)
-except FileExistsError:
-    pass
-
 
 class Controller:
+    """_summary_"""
+
     def __init__(
         self,
-        messenger: Messenger = Messenger(),
-        message_check_delay_s: int = 30,
+        messenger: Messenger,
         routines: Dict[str, Routine] = {},
         monitor: Monitor = Monitor(),
         dispenser: Dispenser = Dispenser(),
-        schedule: AScheduler = AScheduler(),
+        schedule: AdvScheduler = AdvScheduler(),
     ):
         """
         Initialize the Controller.
 
         Parameters
         ----------
-        messenger : Messenger, optional
-            The messenger used for notifications, by default Messenger()
-        message_check_delay_s : int, optional
-            The delay in seconds for checking message commands, by default 30
+        messenger : messenger
+            The messenger used for notifications.
         routines : Dict[str, Routine], optional
             A dictionary of routines, by default {}
         monitor : Monitor, optional
             The monitor for monitoring routines, by default Monitor()
         dispenser : Dispenser, optional
             The dispenser for dispensing, by default Dispenser()
-        schedule : AScheduler, optional
-            The scheduler for routines, by default AScheduler()
+        schedule : AdvScheduler, optional
+            The scheduler for routines, by default AdvScheduler()
         """
         self.messenger = messenger
         self.routines = routines
         self.monitor = monitor
         self.dispenser = dispenser
         self.schedule = schedule
-        self.message_check_delay_s = message_check_delay_s
         self.pending_commands: List[RemoteCommand] = []
 
-        check_job = self.schedule.every(self.message_check_delay_s).seconds
+        check_job = self.schedule.every(self.messenger.check_delay_sec).seconds
         check_job.do(self.check_orders)
+
+    @property
+    def jobs(self) -> List[AdvJob]:
+        """_summary_
+
+        Returns
+        -------
+        List[AdvJob]
+            _description_
+        """
+        return self.schedule.jobs
 
     @property
     def current_schedule(self) -> str:
@@ -85,8 +76,8 @@ class Controller:
         """
         job_strs = []
 
-        for job in self.schedule.jobs:
-            if not (any(["Update " in tag for tag in job.tags])):
+        for job in self.jobs:
+            if not (any(["Update " in str(tag) for tag in job.tags])):
                 job_strs.append(job.to_string(TIME_FMT))
 
         return "\n".join(job_strs)
@@ -110,9 +101,9 @@ class Controller:
             job.priority = routine.priority
             job.tag(name)
             job.tag("Repeating")
-            job.do(self.run_routine, name=name)
+            job.do(self.run, name=name)
 
-    def run_routine(self, name: str) -> Optional[CancelJob]:
+    def run(self, name: str) -> Optional[CancelJob]:
         """
         Run a routine.
 
@@ -127,134 +118,48 @@ class Controller:
             The cancellation job, if applicable.
         """
         routine = self.routines[name]
-
-        start_dt = dt.datetime.now().strftime(TIME_FMT)
-        stop = False
-
-        errors = []
-        run_times: List[float] = []
-        job_ret = None  # Only changed if the job should cancel on errors.
-        error_reports = []
-
-        for step in routine.steps:
-            ret = self.dispenser.run_step(step, self.monitor)
-
-            run_time, errs = ret
-            errors.append(errs)
-            run_times.append(run_time)
-
-            # Case where an initial state error was reported.
-            if run_time == 0:
-                if step.report_invalid_start:
-                    subject = f"{routine.name}: {ret.__class__.__name__}"
-                    body = str(errs[0])
-                    error_reports.append({"subject": subject, "body": body})
-                if not (step.proceed_on_invalid_start):
-                    stop = True
-
-            # Case where the initial state was correctly met
-            elif run_time > 0:
-                # Report and log each individual error encountered.
-                for e in errs:
-                    subject = f"{routine.name}: {e.__class__.__name__}"
-                    body = str(e)
-                    error_reports.append({"subject": subject, "body": body})
-
-                    if isinstance(e, PumpTimeoutError):
-                        if not (step.proceed_on_timeout):
-                            stop = True
-                    elif isinstance(e, ErrorSensorTriggered):
-                        final_run = e.remaining_runs <= 0
-                        if not (step.proceed_on_error) and final_run:
-                            stop = True
-
-            # Figure out if we need to stop we need to stop early...
-            if stop is True:
-                if step.cancel_on_critical_failure:
-                    job_ret = CancelJob()
-                break  # End for loop early
-
-        # Log results and notify user of status.
-        self.log_run(
-            name=routine.name, run_times=run_times, errors=errors, start_dt=start_dt
-        )
-
-        if any(error_reports):
-            for r in error_reports:
-                self.notify(routine.error_contacts, **r)
+        job_ret = routine.run(dispenser=self.dispenser, monitor=self.monitor)
+        for r in routine.error_reports:
+            self.notify(
+                contacts=routine.error_contacts,
+                notify_type=NotifyType.FAILURE,
+                **r,
+            )
 
         if isinstance(job_ret, CancelJob):
-            subject = f"{routine.name}: Critical Error!"
+            title = f"{routine.name}: Critical Error!"
             body = (
                 f"A scheduled job ({routine.name}) was canceled due to a critical "
-                "error!"
+                + "error!"
             )
+            notice_type = NotifyType.FAILURE
         else:
-            subject = f"{routine.name}: Complete!"
-            body = f"A job started at {start_dt} finished running!"
+            title = f"{routine.name}: Complete!"
+            started_at = routine.start_dt.strftime(TIME_FMT)
+            runtime = int(routine.run_time.total_seconds())
+            body = (
+                f"A job started at {started_at} finished running in {runtime} seconds!"
+            )
+            notice_type = NotifyType.SUCCESS
 
-        self.notify(routine.completion_contacts, body, subject)
+        self.notify(
+            body=body,
+            title=title,
+            contacts=routine.completion_contacts,
+            notify_type=notice_type,
+        )
+
+        routine.reset()
 
         return job_ret
 
-    def log_run(
+    def notify(
         self,
-        name: str,
-        run_times: List[float],
-        errors: List[List[Union[bool, Exception]]],
-        start_dt: str,
-    ):
-        """
-        Log the run details of a routine.
-
-        Parameters
-        ----------
-        name : str
-            The name of the routine.
-        run_times : List[float]
-            The list of run times.
-        errors : List[List[Union[bool, Exception]]]
-            The list of errors encountered during the run.
-        start_dt : str
-            The start date and time.
-        """
-        routine_dir = LOG_DIR / name.replace(" ", "_")
-
-        try:
-            routine_dir.mkdir(parents=True, exist_ok=False)
-        except FileExistsError:
-            pass
-
-        logs = zip(run_times, errors)  # Zip these for looping
-
-        for i, log in enumerate(logs):  # Index, run_time, step errors(s)
-            time, errs = log[0], log[1]
-            step_name = self.routines[name].steps[i].name.replace(" ", "_")
-            log_file = routine_dir / f"{step_name}.csv"
-
-            err_dict: Dict[str, List[Any]] = {"timeout": [False]}
-
-            if any(errs):
-                for e in errs:
-                    if isinstance(e, ErrorSensorTriggered):
-                        err_dict[f"{e.name}_error"] = [e.remaining_runs]
-                    if isinstance(e, PumpTimeoutError):
-                        err_dict["timeout"] = [True]
-
-            if time is not None:
-                new_row = {"run_time": [time]}
-                new_row.update(err_dict)
-                new_df = pd.DataFrame(new_row, index=[start_dt])
-
-                if log_file.is_file() and log_file.exists():
-                    existing_df = pd.read_csv(log_file, sep=",", index_col=0)
-                    new_df = pd.concat([existing_df, new_df])
-
-                new_df.to_csv(
-                    log_file, sep=",", index=True, header=True, index_label="Start Time"
-                )
-
-    def notify(self, recipients: List[str], body: str, subject: str = "") -> CancelJob:
+        title: str,
+        body: str,
+        contacts: List[str],
+        notify_type: NotifyType,
+    ) -> CancelJob:
         """
         Send notifications using the Messenger.
 
@@ -276,24 +181,26 @@ class Controller:
         CancelJob
             The cancellation job, if applicable.
         """
-        kwargs: Dict[str, Union[str, List[str]]] = {
-            "body": body,
-            "subject": subject,
-        }
-        if any(recipients):
+        if any(contacts):
             try:
-                self.messenger.send(recipients=recipients, **kwargs)
-            except MessengerError:  # Re-schedules w/ low priority
-                kwargs.update({"recipients": recipients})
+                self.messenger.notify(
+                    body=body,
+                    title=title,
+                    contacts=contacts,
+                    notify_type=notify_type,
+                )
+            except NotificationFailure:
+                params = inspect.signature(self.notify).parameters
+                kwargs = {k: v for k, v in locals().items() if k in params}
                 try_again = self.schedule.every(1).minute
                 try_again.lowest_priority
                 try_again.run_once = True
-                try_again.tag(f"Notify {recipients}")
+                try_again.tag(f"Notify {contacts}")
                 try_again.do(self.notify, **kwargs)
 
-            return CancelJob()
+        return CancelJob()
 
-    def check_orders(self) -> CancelJob:
+    def check_orders(self):
         """
         Check for pending orders using the Messenger.
 
